+++ src/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: tvandivi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/11 12:13:31 by tvandivi          #+#    #+#             */
-/*   Updated: 2019/02/11 12:13:33 by tvandivi         ###   ########.fr       */
+/*   Updated: 2019/02/24 17:57:16 by tvandivi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include <string.h>
+
+size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
